# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TCC II Api",
        "version": "1.5.3"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/dispositivos": {
            "get": {
                "summary": "Exibe dados vindos do Arduino.",
                "operationId": "ad5277a0637d153f4aaf1334d70a4ac8",
                "responses": {
                    "200": {
                        "description": "Há dados para os parâmetros informados"
                    },
                    "400": {
                        "description": "Erro na validação dos campos, seja por campos vazios ou formato inválido"
                    },
                    "404": {
                        "description": "Não há dados para os parâmetros informados"
                    }
                }
            },
            "post": {
                "summary": "Insere dados vindos do Arduino.",
                "operationId": "8a088291b39e16c330c064ca28d7c682",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_arduino": {
                                        "description": "ID do dispositivo do J3M",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "alcool": {
                                        "description": "Total ppm de Álcool",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.686999976634979
                                    },
                                    "benzeno": {
                                        "description": "Total ppm de Benzeno",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.365000009536743
                                    },
                                    "hexano": {
                                        "description": "Total ppm de Hexano",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.153999999165535
                                    },
                                    "metano": {
                                        "description": "Total ppm de Metano",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.365000009536743
                                    },
                                    "fumaca": {
                                        "description": "Total ppm de Fumaça",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.254000008106232
                                    },
                                    "dioxido_carbono": {
                                        "description": "Total ppm de Dióxido de Carbono",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.03599977493286
                                    },
                                    "tolueno": {
                                        "description": "Total ppm de Tolueno",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.498699992895126
                                    },
                                    "amonia": {
                                        "description": "Total ppm de Amônia",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.598699986934662
                                    },
                                    "acetona": {
                                        "description": "Total ppm de Acetona",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    },
                                    "monoxido_carbono": {
                                        "description": "Total ppm de Monóxido de Carbono",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.356000006198883
                                    },
                                    "hidrogenio": {
                                        "description": "Total ppm de Hidrogênio",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.123000003397465
                                    },
                                    "gases_inflamaveis": {
                                        "description": "Total ppm de Gases Inflamáveis",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.119999997317791
                                    },
                                    "temperatura": {
                                        "description": "Temperatura em Graus Celsius",
                                        "type": "number",
                                        "format": "float",
                                        "example": 27.5
                                    },
                                    "umidade": {
                                        "description": "Umidade em percentual",
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Armazenamento dos dados efetuado com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação dos campos, seja por campos vazios ou formato inválido"
                    }
                }
            }
        },
        "/api/users/arduinos": {
            "get": {
                "summary": "Exibe dispositivos cadastrados.",
                "operationId": "84df591735ee684189dc06a2cacf7054",
                "responses": {
                    "200": {
                        "description": "Há dados cadastrados"
                    },
                    "404": {
                        "description": "Não há dados cadastrados"
                    }
                }
            }
        }
    }
}